requirement.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
astroid==2.12.12
asttokens==2.1.0
backcall==0.2.0
black==22.12.0
block==0.0.5
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.2
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==38.0.3
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
Django==3.2.15
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-filter==22.1
django-htmx==1.13.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-types==0.16.0
django-widget-tweaks==1.4.12
executing==1.2.0
fonttools==4.38.0
future==0.18.2
html5lib==1.1
idna==3.4
ipython==8.6.0
isort==5.10.1
jedi==0.18.1
lazy-object-proxy==1.8.0
lxml==4.9.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.4
oauthlib==3.2.2
oscrypto==1.3.0
pandas==1.5.2
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.2
prompt-toolkit==3.0.32
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydyf==0.5.0
Pygments==2.13.0
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyJWT==2.6.0
pylint==2.15.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyPDF3==1.0.6
pyphen==0.13.0
python-barcode==0.14.0
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.0
svglib==1.4.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
wcwidth==0.2.5
weasyprint==57.1
webencodings==0.5.1
wrapt==1.14.1
xhtml2pdf==0.2.8
zopfli==0.2.1
